requirements.txt
python-telegram-bot==13.7
flask==2.0.1
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
python-whois==0.8.0
dnspython==2.4.2
phonenumbers==8.13.27
urllib3==2.1.0
selenium==4.18.1
textblob==0.17.1
nltk==3.8.1
pandas==2.2.1
shodan==1.31.0
censys==2.2.8
python-nmap==0.7.1
cryptography==42.0.5
pyOpenSSL==24.0.0 